+++ src/main/java/adhdmc/nerffarms/command/ReloadCommand.java
@@ -1,6 +1,6 @@
-package adhdmc.nerffarms.Commands;
+package adhdmc.nerffarms.command;
 
-import adhdmc.nerffarms.ConfigParser;
+import adhdmc.nerffarms.config.ConfigParser;
 import adhdmc.nerffarms.NerfFarms;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
 import org.bukkit.command.CommandSender;
